Staging PR for a source code reorganisation (#332)
* add go version to go.mod

* update go.sum for docker checksum change

* move all top-level Go source to `commands` directory in preparation for splitting command builders from `main` package.

* move command builder into a function, out of `main()`

* main.go -> app.go

main.go will go back to the root soon

* changed `commands` into a package and moved the invocation point back into the root `main` package

* remove segment metrics client from codebase

* move version variable (that is injected at compile-time) to `main.go` so the injection works properly
