+++ commands/package_build.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"context"
@@ -7,7 +7,6 @@ import (
 	"runtime"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -58,20 +57,6 @@ func packageBuild(c *cli.Context) error {
 	relativePaths := c.Bool("relativePaths")
 
 	build := c.Args().Get(0)
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package build",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("forceEnsure", forceEnsure).
-				Set("watch", watch).
-				Set("watch", watch).
-				Set("buildFile", buildFile != "").
-				Set("build", build != ""),
-		})
-	}
 
 	cacheDir, err := download.GetCacheDir()
 	if err != nil {
