+++ commands/package_get.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"context"
@@ -6,7 +6,6 @@ import (
 	"os"
 	"strings"
 
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -21,14 +20,6 @@ var packageGetFlags = []cli.Flag{}
 func packageGet(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
-	}
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package get",
-			UserId: config.UserID,
-		})
 	}
 
 	if len(c.Args()) == 0 {
