+++ commands/package_init.go
@@ -1,10 +1,9 @@
-package main
+package commands
 
 import (
 	"context"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -24,14 +23,6 @@ var packageInitFlags = []cli.Flag{
 func packageInit(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
-	}
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package init",
-			UserId: config.UserID,
-		})
 	}
 
 	dir := util.FullPath(c.String("dir"))
