+++ commands/package_release.go
@@ -1,10 +1,9 @@
-package main
+package commands
 
 import (
 	"context"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -24,14 +23,6 @@ var packageReleaseFlags = []cli.Flag{
 func packageRelease(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
-	}
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package release",
-			UserId: config.UserID,
-		})
 	}
 
 	dir := util.FullPath(c.String("dir"))
