+++ commands/package_template_build.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"context"
@@ -6,7 +6,6 @@ import (
 	"path/filepath"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -27,14 +26,6 @@ func packageTemplateBuild(c *cli.Context) (err error) {
 	if len(c.Args()) != 2 {
 		cli.ShowCommandHelpAndExit(c, "build", 0)
 		return nil
-	}
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package template build",
-			UserId: config.UserID,
-		})
 	}
 
 	cacheDir, err := download.GetCacheDir()
