+++ commands/package_template_make.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"context"
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -32,14 +31,6 @@ var packageTemplateMakeFlags = []cli.Flag{
 func packageTemplateMake(c *cli.Context) (err error) {
 	if c.Bool("verbose") {
 		print.SetVerbose()
-	}
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package template make",
-			UserId: config.UserID,
-		})
 	}
 
 	dir := util.FullPath(c.String("dir"))
