+++ commands/package_uninstall.go
@@ -1,13 +1,7 @@
-package main
+package commands
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"strings"
-
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/download"
@@ -17,7 +11,7 @@ import (
 	"github.com/Southclaws/sampctl/versioning"
 )
 
-var packageInstallFlags = []cli.Flag{
+var packageUninstallFlags = []cli.Flag{
 	cli.StringFlag{
 		Name:  "dir",
 		Value: ".",
@@ -25,12 +19,12 @@ var packageInstallFlags = []cli.Flag{
 	},
 	cli.BoolFlag{
 		Name:  "dev",
-		Usage: "for specifying dependencies only necessary for development or testing of the package",
+		Usage: "for specifying development dependencies",
 	},
 }
 
 //nolint:dupl
-func packageInstall(c *cli.Context) error {
+func packageUninstall(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
 	}
@@ -38,18 +32,8 @@ func packageInstall(c *cli.Context) error {
 	dir := util.FullPath(c.String("dir"))
 	development := c.Bool("dev")
 
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package install",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("development", development),
-		})
-	}
-
 	if len(c.Args()) == 0 {
-		cli.ShowCommandHelpAndExit(c, "install", 0)
+		cli.ShowCommandHelpAndExit(c, "uninstall", 0)
 		return nil
 	}
 
@@ -69,33 +53,12 @@ func packageInstall(c *cli.Context) error {
 		return errors.Wrap(err, "failed to interpret directory as Pawn package")
 	}
 
-	err = pcx.Install(context.Background(), deps, development)
+	err = pcx.Uninstall(deps, development)
 	if err != nil {
 		return err
 	}
 
-	print.Info("successfully added new dependency")
+	print.Info("successfully removed dependency")
 
 	return nil
-}
-
-func packageInstallBash(c *cli.Context) {
-	cacheDir, err := download.GetCacheDir()
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "Failed to retrieve cache directory path (attempted <user folder>/.samp) ", err)
-		return
-	}
-
-	packages, err := download.GetPackageList(cacheDir)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "Failed to get package list:", err)
-		return
-	}
-
-	query := c.Args().First()
-	for _, pkg := range packages {
-		if strings.HasPrefix(pkg.String(), query) {
-			fmt.Println(pkg)
-		}
-	}
 }
