+++ commands/server_init.go
@@ -1,10 +1,9 @@
-package main
+package commands
 
 import (
 	appRuntime "runtime"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
 	"gopkg.in/urfave/cli.v1"
 
 	"github.com/Southclaws/sampctl/print"
@@ -32,16 +31,6 @@ func serverInit(c *cli.Context) error {
 
 	version := c.String("version")
 	dir := util.FullPath(c.String("dir"))
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "server init",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("version", version),
-		})
-	}
 
 	err := runtime.InitialiseServer(version, dir, appRuntime.GOOS)
 	if err != nil {
